Version 0.9.1
* Fix issue when snapshot was detected in file extension
* Change buttons text
* Disable delete buttons if action is not performed
* Disable tree item edit
* Add option for autoclean of dist folder
* Use type alias PathArray
* Add files for binary build with Docker image
* Add option for autoclean of dist folder (.bat file)
* Change term binary -> regular
* Update README.md
* Update shebang
* Add instructions for native build
* Check empty list of TARs
* Add instructions for building
* Introduce BIN_NAME variable
* Add open file function
* Fix quotes
* Remove build folder and .spec file
* Introduce variable for binary name
* Exclude archive folder
* Move types definition
* Add class for filtering files by snapshot
* Use Tester instance
* Rename method & field
* Update type hints

+++ gui/ui.py
@@ -1,5 +1,6 @@
 import os
 import shutil
+import subprocess
 
 from PyQt5 import QtWidgets, uic, QtGui
 from PyQt5.QtCore import QModelIndex, QThreadPool, Qt
@@ -7,7 +8,8 @@ from PyQt5.QtGui import QDragEnterEvent, QDropEvent
 from PyQt5.QtWidgets import QFileDialog, QMenu
 
 import core.file
-from core.tree import TreeType, OperationType, FileTreeWorker
+from core.tree import FileTreeWorker
+from core.types import TreeType, OperationType
 from core.workers import ClearSnapshotWorker, ClearEmptyDirsWorker
 from gui import icons
 
@@ -29,8 +31,9 @@ class HistoryUI(QtWidgets.QMainWindow):
         icons.IconsLoader(project_home)
 
         # Connect signals
-        FileTreeWorker.signals.finished.connect(self.update_tree)      # Connect to slot for finishing
-        FileTreeWorker.signals.finished.connect(self.switch_clear_all) # Switch Clear all button
+        FileTreeWorker.signals.finished.connect(self.update_tree)               # Connect to slot for finishing
+        FileTreeWorker.signals.finished.connect(self.switch_clear_all)          # Switch Clear all button
+        FileTreeWorker.signals.finished.connect(self.switch_delete_snapshots)   # Switch Clear all button
 
         ClearSnapshotWorker.signals.finished.connect(lambda: self.statusbar.showMessage("Snapshots are cleared"))
         ClearEmptyDirsWorker.signals.finished.connect(lambda: self.statusbar.showMessage("Empty directories are cleared"))
@@ -83,6 +86,13 @@ class HistoryUI(QtWidgets.QMainWindow):
             index = index.parent()
         return selected_path, selected_item
 
+    def _open_file(self):
+        file = self.get_selected_path()[0]
+        try:
+            os.startfile(file)
+        except AttributeError:
+            subprocess.call(['open', file])
+
     def browse_action(self) -> None:
         dialog = QFileDialog(self)
         dialog.setFileMode(QFileDialog.FileMode.Directory)
@@ -98,6 +108,9 @@ class HistoryUI(QtWidgets.QMainWindow):
     def switch_clear_all(self, op_type: OperationType, _) -> None:
         self.clear_all_button.setEnabled(op_type == OperationType.EMPTY_DIRS)
 
+    def switch_delete_snapshots(self, op_type: OperationType, _) -> None:
+        self.delete_button.setEnabled(op_type == OperationType.FILTERED_TREE)
+
     def build_file_tree(self, op_type: OperationType) -> None:
         if self.path_field.text():
             worker = FileTreeWorker(self.path_field.text(),
@@ -106,7 +119,9 @@ class HistoryUI(QtWidgets.QMainWindow):
 
             # Switch filter
             if op_type == OperationType.FILTERED_TREE:
-                worker.set_filter(self.filter_from_field.text(), self.filter_to_field.text())
+                tester = core.file.SnapshotTester([self.filter_from_field.text(), self.filter_to_field.text()],
+                                                  self.checked()[0])
+                worker.set_filter(tester)
 
             # Start worker in another thread
             QThreadPool.globalInstance().start(worker)
@@ -158,19 +173,9 @@ class HistoryUI(QtWidgets.QMainWindow):
 
     def delete_snapshots_action(self) -> None:
         if self.path_field.text():
-            from_snapshot = self.filter_from_field.text()
-            to_snapshot = self.filter_to_field.text()
-
-            if self.checked()[0] == TreeType.UNIFIED:
-                snapshot_fun = lambda root, file: core.file.get_snapshot(file)
-            if self.checked()[0] == TreeType.BY_DATE:
-                snapshot_fun = lambda root, file: root.split(os.sep)[1]
-
-            def test_fun(root: str, file: str) -> bool:
-                snapshot = snapshot_fun(root, file)
-                return (from_snapshot <= snapshot) and (snapshot <= to_snapshot)
-
-            worker = ClearSnapshotWorker(self.path_field.text(), test_fun)
+            bounds = [self.filter_from_field.text(), self.filter_to_field.text()]
+            tester = core.file.SnapshotTester(bounds, self.checked()[0])
+            worker = ClearSnapshotWorker(self.path_field.text(), tester)
             worker.signals.progress.connect(lambda x: print(x))
             QThreadPool.globalInstance().start(worker)
             self.statusbar.showMessage("Start clear snapshots")
@@ -194,35 +199,46 @@ class HistoryUI(QtWidgets.QMainWindow):
 
     def contextMenuEvent(self, position) -> None:
         # Don't open context menu for no selection
-        if len(self.get_selected_nodes()) == 0:
+        nodes = self.get_selected_nodes()
+        if len(nodes) == 0:
             return
 
         context_menu = QMenu()
-        restore = context_menu.addAction("Restore")
-        context_menu.addSeparator()
-        from_snapshot = context_menu.addAction("From snapshot")
-        to_snapshot = context_menu.addAction("To snapshot")
-        context_menu.addSeparator()
-        expand = context_menu.addAction("Expand recursively")
-        context_menu.addSeparator()
-        delete = context_menu.addAction("Delete empty directory")
-        action = context_menu.exec_(self.file_tree_view.mapToGlobal(position))
-        if action == restore:
-            self.restore_action()
-        if action == from_snapshot:
-            selected_nodes = self.get_selected_nodes()
-            snapshot = selected_nodes[0].siblingAtColumn(2).data()
-            self.filter_from_field.setText(snapshot)
-        if action == to_snapshot:
-            selected_nodes = self.get_selected_nodes()
-            snapshot = selected_nodes[0].siblingAtColumn(2).data()
-            self.filter_to_field.setText(snapshot)
-        if action == expand:
-            self.file_tree_view.expandRecursively(self.get_selected_nodes()[0])
-        if action == delete:
-            selected_path, _ = self.get_selected_path()
-            try:
-                os.removedirs(selected_path)
-                self.statusbar.showMessage("Delete", selected_path)
-            except OSError:
-                self.statusbar.showMessage("Failed to delete", selected_path)
+        if nodes[0].siblingAtColumn(1).data() != "Folder":
+            openfile = context_menu.addAction("Open")
+            restore = context_menu.addAction("Restore")
+            context_menu.addSeparator()
+            from_snapshot = context_menu.addAction("From snapshot")
+            to_snapshot = context_menu.addAction("To snapshot")
+            action = context_menu.exec_(self.file_tree_view.mapToGlobal(position))
+
+            if action == openfile:
+                self._open_file()
+            if action == restore:
+                self.restore_action()
+            if action == from_snapshot:
+                selected_nodes = self.get_selected_nodes()
+                snapshot = selected_nodes[0].siblingAtColumn(2).data()
+                self.filter_from_field.setText(snapshot)
+            if action == to_snapshot:
+                selected_nodes = self.get_selected_nodes()
+                snapshot = selected_nodes[0].siblingAtColumn(2).data()
+                self.filter_to_field.setText(snapshot)
+        else:
+            expand = context_menu.addAction("Expand recursively")
+
+            delete = None
+            if self.clear_all_button.isEnabled():
+                context_menu.addSeparator()
+                delete = context_menu.addAction("Delete empty directory")
+
+            action = context_menu.exec_(self.file_tree_view.mapToGlobal(position))
+            if action == expand:
+                self.file_tree_view.expandRecursively(self.get_selected_nodes()[0])
+            if self.clear_all_button.isEnabled() and action == delete:
+                selected_path, _ = self.get_selected_path()
+                try:
+                    os.removedirs(selected_path)
+                    self.statusbar.showMessage("Delete", selected_path)
+                except OSError:
+                    self.statusbar.showMessage("Failed to delete", selected_path)
